Added aqua to crosshair color rotation and removed turn bind
I've been using aqua from time to time lately, so I figure it makes sense to have it in the rotation.  I don't ever use the turn bind so I removed it.  I might put the turn bind back in at some point if I find some use for it.